Refactored MainScreen
